requirements.txt
# Django
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.5,<24.0

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# Environment
python-dotenv>=1.0,<2.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Security
argon2-cffi>=23.1,<24.0

# File Storage
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
Pillow>=10.2,<11.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0

# Development
ipython>=8.20,<9.0
